SIDE-backend/requirements.txt
azure-core==1.24.0
Brotli==1.0.9
camelot-py==0.10.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
cycler==0.11.0
dash==2.4.1
dash-bootstrap-components==1.1.0
dash-colorscales==0.0.4
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
distlib==0.3.4
distro==1.7.0
dtale==2.4.0
Elixir==0.7.1
et-xmlfile==1.1.0
filelock==3.4.2
Flask==2.1.1
Flask-Compress==1.12
Flask-Cors==3.0.10
flask-ngrok==0.0.25
fonttools==4.28.5
future==0.18.2
ghostscript==0.7
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
kaleido==0.2.1
kiwisolver==1.3.2
lz4==4.0.0
MarkupSafe==2.1.1
matplotlib==3.5.2
missingno==0.4.2
networkx==2.8
numpy==1.21.2
opencv-python==4.5.4.60
openpyxl==3.0.9
packaging==21.3
pandas==1.3.3
patsy==0.5.2
pbr==5.9.0
pdfminer.six==20220506
pdftopng==0.2.3
Pillow==8.4.0
platformdirs==2.4.1
plotly==5.8.0
pycparser==2.21
pymongo==4.0.1
pyparsing==3.0.6
PyPDF2==1.27.12
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
scikit-learn==1.1.0
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.4.2
squarify==0.4.3
statsmodels==0.13.2
strsimpy==0.2.1
tabula-py==2.3.0
tabulate==0.8.9
Tempita==0.5.2
tenacity==8.0.1
threadpoolctl==3.1.0
typing-extensions==4.2.0
urllib3==1.26.9
virtualenv==20.13.0
Werkzeug==2.1.1
xarray==2022.3.0
xlrd==0.7.1
xlwt==0.7.2
zipp==3.8.0
